refactor sprites, implement test
changed the center of all sprites (except missiles) to (0,0).  This
facilitates tile-based movement and map drawing.  Implemented a test
case of a small room with 1 pc and 1 npc.  NPC has simple AI to move to
the player, but the implementation is buggy due to user input handling
not being integrated with game engine FSM.